backend/requirements.txt
fastapi>=0.115.2
uvicorn>=0.30.0
pydantic>=2.10.3
pydantic[email]>=2.10.3
pydantic-settings>=2.10.1
email-validator>=2.1.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
alembic>=1.13.1
# AI / RAG Dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.26.0
mistralai>=0.0.10
python-json-logger>=2.0.7
# Optional: llama-cpp-python for local LLM (can be tricky on Windows, user might need Visual Studio Build Tools)
# llama-cpp-python
